requirements.txt
Pillow==10.1.0
# torchとtorchvisionはプラットフォーム別にインストールが必要です
# - Windows (CUDA 12.1): pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# - Windows (CUDA 12.4): pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
# - Windows (CUDA 13.1): CUDA 12.4版を使用（後方互換性あり）
#   → pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
torch>=2.1.0
torchvision>=0.16.0
transformers==4.40.0
sentencepiece==0.1.99
accelerate==0.30.1
sse-starlette==2.1.0
fastapi==0.111.0
numpy==1.26.4
PyYAML>=6.0
pynvml>=11.5.0
bitsandbytes>=0.41.0

# Testing dependencies
pytest
pytest-asyncio
pytest-cov
httpx
